backend/requirement.txt
asgiref==3.4.1
autopep8==1.6.0
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
pycodestyle==2.8.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5

numpy==1.21.6
python-dotenv==1.0.0
openai==0.27.0
langchain==0.0.221
langchain-openai==0.0.2
langchain-community==0.0.1

chroma-db==0.3.21